+++ config/services.php
@@ -15,6 +15,7 @@ use Ajgarlag\Bundle\OpenIDConnectProviderBundle\Logout\PostLogoutRedirectUriStor
 use Ajgarlag\Bundle\OpenIDConnectProviderBundle\Manager\RelyingPartyManagerInterface;
 use Ajgarlag\Bundle\OpenIDConnectProviderBundle\OAuth2\IdTokenGrant;
 use Ajgarlag\Bundle\OpenIDConnectProviderBundle\OpenIDConnect\IdTokenResponse;
+use Ajgarlag\Bundle\OpenIDConnectProviderBundle\OpenIDConnect\SessionSidManager;
 use Ajgarlag\Bundle\OpenIDConnectProviderBundle\Repository\IdentityProvider;
 use League\Bundle\OAuth2ServerBundle\Manager\ClientManagerInterface;
 use OpenIDConnectServer\ClaimExtractor;
@@ -33,12 +34,19 @@ return static function (ContainerConfigurator $container): void {
 
         ->set('ajgarlag.openid_connect_provider.openid_connect.claim_extractor', ClaimExtractor::class)
 
+        ->set('ajgarlag.openid_connect_provider.openid_connect.session_sid_manager', SessionSidManager::class)
+            ->args([
+                service('request_stack'),
+                service('security.helper'),
+            ])
+
         ->set('ajgarlag.openid_connect_provider.openid_connect.response', IdTokenResponse::class)
             ->args([
                 service('ajgarlag.openid_connect_provider.repository.identity_provider'),
                 service('ajgarlag.openid_connect_provider.openid_connect.claim_extractor'),
                 service('event_dispatcher'),
                 service('request_stack'),
+                service('ajgarlag.openid_connect_provider.openid_connect.session_sid_manager'),
             ])
         ->alias(IdTokenResponse::class, 'ajgarlag.openid_connect_provider.openid_connect.response')
 
@@ -93,6 +101,7 @@ return static function (ContainerConfigurator $container): void {
                 null,
                 service(PostLogoutRedirectUriStorageInterface::class),
                 service('security.helper'),
+                service('ajgarlag.openid_connect_provider.openid_connect.session_sid_manager'),
                 service('twig'),
                 service('security.http_utils'),
                 null,